Start on workflow implementation, starting with the first workflow: Going to the writing phase.
This commit also implements the real-time updates for the workflow updates via the RetrospectiveStatusUpdateDispatcher.
This commit also lays the groundwork for the rest of the workflow, which from the UI are driven through the Panels.

Some helper components like RetrospectiveStagePanel, RetrospectiveStageTitle, and RetrospectiveStatusPanel were added which react in rendering to their cascaded values.