Optimizing the test execution time by cutting down on setup and teardown
We previously started and stopped the web server on every fixture, in addition to setting up and tearing down the data on every fixture. Not every test case is still isolated, but they share the expensive database and server set-up. This also has the advantage that it simulates a more app-like load. Web drivers are also shared in a WebDriverPool.

Before optimization: 3,88 minutes
After optimization: 3,01 minutes